+++ src/ts/markdown/md2html.ts
@@ -19,6 +19,7 @@ export const loadLuteJs = (vditor: IVditor | string) => {
         vditor.lute.SetInlineMathAllowDigitAfterOpenMarker(vditor.options.preview.math.inlineDigit);
         vditor.lute.SetAutoSpace(vditor.options.preview.markdown.autoSpace);
         vditor.lute.SetToC(vditor.options.preview.markdown.toc);
+        vditor.lute.SetFootnotes(vditor.options.preview.markdown.footnotes);
         vditor.lute.SetChinesePunct(vditor.options.preview.markdown.chinesePunct);
         vditor.lute.SetFixTermTypo(vditor.options.preview.markdown.fixTermTypo);
     }
@@ -41,6 +42,7 @@ export const md2htmlByPreview = (mdText: string, options?: IPreviewOptions) => {
     lute.SetInlineMathAllowDigitAfterOpenMarker(options.math.inlineDigit);
     lute.SetAutoSpace(options.markdown.autoSpace);
     lute.SetToC(options.markdown.toc);
+    lute.SetFootnotes(options.markdown.footnotes);
     lute.SetChinesePunct(options.markdown.chinesePunct);
     lute.SetFixTermTypo(options.markdown.fixTermTypo);
     return lute.Md2HTML(mdText);